+++ code/src/App.js
@@ -1,10 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import moment from 'moment';
 
-import { THOUGHTS_URL } from './urls';
+import { ThoughtList } from 'components/ThoughtList';
 import { ThoughtsInput } from './components/ThoughtsInput';
-import { Thought } from './components/Thought';
-import { ThoughtLikes } from './components/ThoughtLikes';
+import { THOUGHTS_URL } from './urls';
 
 export const App = () => {
   const [thoughts, setThoughts] = useState([]);
@@ -26,15 +24,6 @@ export const App = () => {
       .catch((error) => console.error(error));
   }
 
-  const onLikedThought = (id) => {
-    fetch(`THOUGHTS_URL/${id}/like`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: ''
-    })
-      .then(() => ThoughtLikes(id))
-  }
-
   useEffect(() => {
     fetchThoughts();
   }, []);
@@ -42,24 +31,7 @@ export const App = () => {
   return (
     <div>
       <ThoughtsInput onThoughtChange={postThoughts} />
-      <div>
-        {thoughts.map((thought) => {
-          return (
-            <Thought
-              message={thought.message}
-              hearts={thought.hearts}
-              time={moment(thought.createdAt).fromNow()}
-              key={thought._id}
-              id={thought._id}
-              setThoughts={setThoughts}
-              thoughts={thoughts}
-              onLikedThought={onLikedThought} />
-          )
-        })}
-      </div>
-      <div>
-        <ThoughtLikes onLikedThought={onLikedThought} />
-      </div>
+      <ThoughtList thoughts={thoughts} setThoughts={setThoughts} />
     </div>
   );
 };
