+++ code/src/components/Thought.js
@@ -5,17 +5,18 @@ export const Thought = ({
   message,
   hearts,
   time,
-  id
+  id,
+  onLikedThought
 }) => {
   return (
     <div>
       <p key={id} id={id}>
         {message}
         <span className="thought-time">{time}</span>
-        <div className="liked-container">
-          <ThoughtLikes id={id} />
+        <span className="liked-container">
+          <ThoughtLikes id={id} onLikedThought={onLikedThought} />
           <span className="liked"> x {hearts}</span>
-        </div>
+        </span>
       </p>
     </div>
   );
