+++ code/src/components/ThoughtLikes.js
@@ -1,24 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-export const ThoughtLikes = ({ setThoughts, thoughts }) => {
-  const [likedThought, setLikedThougt] = useState([]);
-  const getLikedThought = (id) => {
-    const isLiked = thoughts.map((thought) => {
-      if (id === thought._id) {
-        thought.hearts += 1;
-      }
-      return thought;
-    });
-    setThoughts(isLiked);
-    if (!likedThought.includes(id)) {
-      setLikedThougt([...likedThought, id]);
-    }
-  };
-
+export const ThoughtLikes = ({ id, onLikedThought }) => {
   return (
-    <div>
+    <button type="button" onClick={() => onLikedThought(id)}>
       <span role="img" aria-label="red heart">❤️</span>
-      <p>{getLikedThought}</p>
-    </div>
+    </button>
   )
 };
